Config and change respone message
