+++ socket-service/src/main/java/com/hcmus/socketservice/entity/dto/Message.java
@@ -1,4 +1,4 @@
-package com.hcmus.socketservice.message;
+package com.hcmus.socketservice.entity.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -6,14 +6,14 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 
 /**
- * Message sent to client
+ * Message sent to server by browser
  *
- * 封装发给客户端的消息
+ * 封装浏览器发送到服务器的消息
  */
 @Getter
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class Response {
-    private String response;
+public class Message {
+    private String name;
 }
