+++ socket-service/src/main/java/com/hcmus/socketservice/entity/dto/Response.java
@@ -1,0 +1,21 @@
+package com.hcmus.socketservice.entity.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+/**
+ * Message sent to client
+ *
+ * 封装发给客户端的消息
+ */
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Response {
+    private String response;
+    private Double latPickingAddress;
+    private Double lngPickingAddress;
+}
