--- socket-service/src/main/java/com/hcmus/socketservice/message/Message.java
@@ -1,19 +1,0 @@
-package com.hcmus.socketservice.message;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-/**
- * Message sent to server by browser
- *
- * 封装浏览器发送到服务器的消息
- */
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Message {
-    private String name;
-}
