+++ socket-service/src/main/resources/application.yaml
@@ -1,2 +1,6 @@
+spring:
+  config:
+    import: optional:secret.yaml
+
 server:
   port: 8082
